[UI] Доработаны формы авторизации/регистрации Resolve #569
