+++ src/frontend/components/logo/yandex.tsx
@@ -1,0 +1,21 @@
+export const Yandex = ({ className }: { className: string }) => {
+  return (
+    <svg
+      className={className}
+      fill="currentColor"
+      height="1em"
+      stroke="currentColor"
+      strokeWidth="0"
+      viewBox="0 0 44 45"
+      width="1em"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path d="M0 0h24v24H0z" fill="none" />
+      <path
+        d="M31.1666 37.1663H26.3485V11.7978H24.1969C20.2576 11.7978 18.1969 13.8931 18.1969 
+      17.0198C18.1969 20.5656 19.6212 22.2096 22.5606 24.3048L24.9848 26.0455L18.0151 
+      37.1663H12.8333L19.106 27.2381C15.5 24.4982 13.4697 21.8228 13.4697 17.3099C13.4697 11.6689 17.1666 7.83301 24.1666 7.83301H31.1363L31.1666 37.1663Z"
+      />
+    </svg>
+  );
+};
