+++ src/frontend/lib/models/wish.ts
