+++ src/frontend/lib/models/wishlist.ts
